Restructure the files and code and change the names of images
