+++ public/sidebar/currency-dollar-white.png
